# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: 2pick migration API
  version: 0.1.0
  description: >-
    Contract shared by the Laravel identity bridge and the versioned Go API.
    Existing /api routes remain owned by Laravel until their individual cutover.
servers:
  - url: https://2pick.app
  - url: http://localhost
...
paths:
...
  /api/v1/auth/me:
    get:
      operationId: getGoIdentity
      tags: [Go auth]
      summary: Verify a token and return its identity, with the account behind it
      description: >
        The `user` object replaces the `user` half of Laravel's /session-context, which is
        what the SPA read to draw its own account. Present whenever this process can reach
        the database; without one the endpoint stays a pure token check. has_password and
        linked_google drive the profile page: 82% of accounts have no password, and
        offering them "change password" instead of "set a password" is the difference
        between a working form and one that can never succeed.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The token's claims, plus the account they name when available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityEnvelope"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/auth/login:
    post:
      operationId: goLogin
      tags: [Go auth]
      summary: Exchange an address and password for a session
      description: >
        The access token is in the body; the refresh token is NOT — it is set as an
        httpOnly cookie so script cannot read it, and the CSRF value the client must
        echo back is set as a readable cookie alongside it. Every failure answers 401
        with the same code, deliberately: distinguishing an unknown address from a wrong
        password turns this into an account enumeration oracle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, maxLength: 255 }
                password: { type: string, maxLength: 72 }
      responses:
        "200":
          description: A new session. Sets the 2pick_refresh and 2pick_csrf cookies.
          headers:
            Set-Cookie:
              description: 2pick_refresh (httpOnly) and 2pick_csrf (readable)
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrantEnvelope"
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/auth/register:
    post:
      operationId: goRegister
      tags: [Go auth]
      summary: Create a password account and sign it in
      description: >
        Moved off Laravel because it needed no mail: the User model does not implement
        MustVerifyEmail, so registration was only ever a validator, an insert and a
        login. Signing in on success matches what the RegistersUsers trait did.

        A refused sign-up answers 422 with per-field machine codes in `data.errors`
        (required, too_long, too_short, invalid_email, taken, mismatch) rather than with
        sentences: this API has no message catalogue and the client serves three
        languages.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, password_confirmation]
              properties:
                name: { type: string, maxLength: 20, description: Counted in runes, not bytes }
                email: { type: string, maxLength: 50 }
                password: { type: string, minLength: 8, maxLength: 72 }
                password_confirmation: { type: string }
      responses:
        "200":
          description: The account was created and is signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrantEnvelope"
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "422":
          description: Per-field validation codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorsEnvelope"
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/auth/refresh:
    post:
      operationId: goRefreshSession
      tags: [Go auth]
      summary: Rotate the session
      description: >
        Reads the refresh token from the 2pick_refresh cookie and requires the CSRF value
        in X-CSRF-Token. The presented token is consumed and a new one issued in the same
        family; presenting a token that was already consumed revokes the whole family,
        because there is no way to tell the attacker's copy from the victim's.
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A rotated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrantEnvelope"
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/auth/logout:
    post:
      operationId: goLogout
      tags: [Go auth]
      summary: Revoke the session family
      description: >
        Requires X-CSRF-Token when a session cookie is present. Succeeds even when the
        token is already unknown to the server, so a client can always escape a broken
        session.
      parameters:
        - in: header
          name: X-CSRF-Token
          schema: { type: string }
      responses:
        "204": { description: The session is gone and the cookies are cleared }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/auth/oauth/google/start:
    get:
      operationId: startGoogleSignIn
      tags: [Go auth]
      summary: Redirect the browser to Google
      description: >
        A 302, because the browser navigates here directly. return_to is checked against
        the allowed origins when the flow starts and remembered server-side; a target
        outside the list is replaced with the default rather than refused.
      parameters:
        - in: query
          name: return_to
          schema: { type: string }
      responses:
        "302": { description: Redirect to Google's consent screen }
        "500": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/auth/oauth/google/callback:
    get:
      operationId: completeGoogleSignIn
      tags: [Go auth]
      summary: Finish the Google flow
      description: >
        EVERY outcome is a 302, including the failures: this is reached by a top-level
        navigation, so the caller is looking at a page rather than at a JSON body. The
        result is reported as ?auth=signed-in|registered|linked|failed, with ?reason= on
        a failure. No token ever appears in the redirect URL — the session arrives as
        cookies on this response, and the client calls /api/v1/auth/refresh to obtain an
        access token.
      parameters:
        - { in: query, name: state, schema: { type: string } }
        - { in: query, name: code, schema: { type: string } }
        - { in: query, name: error, schema: { type: string } }
      responses:
        "302":
          description: Redirect back to the SPA, with the session cookies on success
          headers:
            Set-Cookie:
              description: Set on a login, cleared on a failure, absent on a link
              schema: { type: string }
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/auth/oauth/google/connect:
    post:
      operationId: connectGoogleAccount
      tags: [Go auth]
      summary: Start linking a Google account to the caller's account
      description: >
        Answers with a URL instead of redirecting, and is POST rather than GET, because
        linking has to know who is asking — which needs an Authorization header, and a
        browser cannot put one on a top-level navigation. The client reads the URL and
        navigates itself.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: return_to
          schema: { type: string }
      responses:
        "200":
          description: The URL to navigate to
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorization_url: { type: string }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "500": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
  /api/v1/posts:
    get:
      operationId: listPublicPosts
      tags: [Go public content]
      parameters:
        - { in: query, name: sort_by, schema: { type: string, enum: [hot, new], default: hot } }
        - { in: query, name: range, schema: { type: string, enum: [all, year, month, week, day], default: week } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: per_page, schema: { type: integer, minimum: 1, maximum: 15, default: 15 } }
        - { in: query, name: k, schema: { type: string, maxLength: 255 } }
      responses:
        "200":
          description: Paginated public voting cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsEnvelope"
        "422": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
...
  /api/v1/tags:
    get:
      operationId: listPublicTags
      tags: [Go public content]
      parameters:
        - { in: query, name: keyword, schema: { type: string, maxLength: 30 } }
      responses:
        "200":
          description: Tags used by public posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagsEnvelope"
        "422": { $ref: "#/components/responses/ErrorResponse" }
        "503": { $ref: "#/components/responses/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Ed25519 JWT
...
  responses:
    ErrorResponse:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    IdentityEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [subject, roles, expires_at]
              properties:
                subject:
                  type: string
                roles:
                  type: array
                  items: { type: string }
                expires_at:
                  type: string
                  format: date-time
                user:
                  description: >
                    Absent when this process cannot reach the database. Replaces the
                    `user` half of Laravel's /session-context.
                  type: object
                  required: [user_id, name, avatar_url, has_password, linked_google]
                  properties:
                    user_id: { type: string }
                    name: { type: string }
                    avatar_url: { type: [string, "null"] }
                    has_password:
                      type: boolean
                      description: False for the 82% of accounts that signed up through Google.
                    linked_google: { type: boolean }
...